Refactor (Path|PortableDataStream)DataSetIterator
