+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/iterator/PathSparkDataSetIterator.java
@@ -21,18 +21,9 @@ import java.util.List;
  *
  * @author Alex Black
  */
-public class PathSparkDataSetIterator implements DataSetIterator {
+public class PathSparkDataSetIterator extends BaseDataSetIterator<String> {
 
     public static final int BUFFER_SIZE = 4194304;  //4 MB
-
-    private final Collection<String> dataSetStreams;
-    private DataSetPreProcessor preprocessor;
-    private Iterator<String> iter;
-    private int totalOutcomes = -1;
-    private int inputColumns = -1;
-    private int batch = -1;
-    private int cursor = 0;
-    private DataSet preloadedDataSet;
     private FileSystem fileSystem;
 
     public PathSparkDataSetIterator(Iterator<String> iter){
@@ -46,78 +37,8 @@ public class PathSparkDataSetIterator implements DataSetIterator {
     }
 
     @Override
-    public DataSet next(int num) {
-        return next();
-    }
-
-    @Override
     public int totalExamples() {
         throw new UnsupportedOperationException("Total examples unknown for PathSparkDataSetIterator");
-    }
-
-    @Override
-    public int inputColumns() {
-        if(inputColumns == -1) preloadDataSet();
-        return inputColumns;
-    }
-
-    @Override
-    public int totalOutcomes() {
-        if(totalOutcomes == -1) preloadDataSet();
-        return totalExamples();
-    }
-
-    @Override
-    public boolean resetSupported(){
-        return dataSetStreams != null;
-    }
-
-    @Override
-    public boolean asyncSupported() {
-        return true;
-    }
-
-    @Override
-    public void reset() {
-        if(dataSetStreams == null) throw new IllegalStateException("Cannot reset iterator constructed with an iterator");
-        iter = dataSetStreams.iterator();
-        cursor = 0;
-    }
-
-    @Override
-    public int batch() {
-        if(batch == -1) preloadDataSet();
-        return batch;
-    }
-
-    @Override
-    public int cursor() {
-        return cursor;
-    }
-
-    @Override
-    public int numExamples() {
-        return 0;
-    }
-
-    @Override
-    public void setPreProcessor(DataSetPreProcessor preProcessor) {
-        this.preprocessor = preProcessor;
-    }
-
-    @Override
-    public DataSetPreProcessor getPreProcessor() {
-        return this.preprocessor;
-    }
-
-    @Override
-    public List<String> getLabels() {
-        return null;
-    }
-
-    @Override
-    public boolean hasNext() {
-        return iter.hasNext();
     }
 
     @Override
@@ -138,19 +59,7 @@ public class PathSparkDataSetIterator implements DataSetIterator {
         return ds;
     }
 
-    @Override
-    public void remove() {
-        throw new UnsupportedOperationException();
-    }
-
-    private void preloadDataSet(){
-        preloadedDataSet = load(iter.next());
-        totalOutcomes = preloadedDataSet.getLabels().size(1);
-        inputColumns = preloadedDataSet.getFeatureMatrix().size(1);
-        batch = preloadedDataSet.numExamples();
-    }
-
-    private synchronized DataSet load(String path){
+    protected synchronized DataSet load(String path){
         if(fileSystem == null){
             try{
                 fileSystem = FileSystem.get(new URI(path), new Configuration());
