+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/iterator/PortableDataStreamDataSetIterator.java
@@ -18,16 +18,7 @@ import java.util.List;
  *
  * @author Alex Black
  */
-public class PortableDataStreamDataSetIterator implements DataSetIterator {
-
-    private final Collection<PortableDataStream> dataSetStreams;
-    private DataSetPreProcessor preprocessor;
-    private Iterator<PortableDataStream> iter;
-    private int totalOutcomes = -1;
-    private int inputColumns = -1;
-    private int batch = -1;
-    private int cursor = 0;
-    private DataSet preloadedDataSet;
+public class PortableDataStreamDataSetIterator extends BaseDataSetIterator<PortableDataStream> {
 
     public PortableDataStreamDataSetIterator(Iterator<PortableDataStream> iter){
         this.dataSetStreams = null;
@@ -40,78 +31,8 @@ public class PortableDataStreamDataSetIterator implements DataSetIterator {
     }
 
     @Override
-    public DataSet next(int num) {
-        return next();
-    }
-
-    @Override
     public int totalExamples() {
         throw new UnsupportedOperationException("Total examples unknown for PortableDataStreamDataSetIterator");
-    }
-
-    @Override
-    public int inputColumns() {
-        if(inputColumns == -1) preloadDataSet();
-        return inputColumns;
-    }
-
-    @Override
-    public int totalOutcomes() {
-        if(totalOutcomes == -1) preloadDataSet();
-        return totalExamples();
-    }
-
-    @Override
-    public boolean resetSupported(){
-        return dataSetStreams != null;
-    }
-
-    @Override
-    public boolean asyncSupported() {
-        return true;
-    }
-
-    @Override
-    public void reset() {
-        if(dataSetStreams == null) throw new IllegalStateException("Cannot reset iterator constructed with an iterator");
-        iter = dataSetStreams.iterator();
-        cursor = 0;
-    }
-
-    @Override
-    public int batch() {
-        if(batch == -1) preloadDataSet();
-        return batch;
-    }
-
-    @Override
-    public int cursor() {
-        return cursor;
-    }
-
-    @Override
-    public int numExamples() {
-        return 0;
-    }
-
-    @Override
-    public void setPreProcessor(DataSetPreProcessor preProcessor) {
-        this.preprocessor = preProcessor;
-    }
-
-    @Override
-    public DataSetPreProcessor getPreProcessor() {
-        return this.preprocessor;
-    }
-
-    @Override
-    public List<String> getLabels() {
-        return null;
-    }
-
-    @Override
-    public boolean hasNext() {
-        return iter.hasNext();
     }
 
     @Override
@@ -132,19 +53,7 @@ public class PortableDataStreamDataSetIterator implements DataSetIterator {
         return ds;
     }
 
-    @Override
-    public void remove() {
-        throw new UnsupportedOperationException();
-    }
-
-    private void preloadDataSet(){
-        preloadedDataSet = load(iter.next());
-        totalOutcomes = preloadedDataSet.getLabels().size(1);
-        inputColumns = preloadedDataSet.getFeatureMatrix().size(1);
-        batch = preloadedDataSet.numExamples();
-    }
-
-    private DataSet load(PortableDataStream pds){
+    protected DataSet load(PortableDataStream pds){
         DataSet ds = new DataSet();
         try(InputStream is = pds.open()){
             ds.load(is);
@@ -154,4 +63,5 @@ public class PortableDataStreamDataSetIterator implements DataSetIterator {
         cursor++;
         return ds;
     }
+
 }
